OV-298: * types
