+++ frontend/src/bundles/studio/types/script-with-icon.type.ts
@@ -1,0 +1,8 @@
+import { type ValueOf } from '~/bundles/common/types/types.js';
+
+import { type PlayIconNames } from '../enums/enums.js';
+import { type Script } from './types.js';
+
+type ScriptWithIcon = Script & { iconName: ValueOf<typeof PlayIconNames> };
+
+export { type ScriptWithIcon };
