+++ shared/src/bundles/avatar-videos/avatar-videos.ts
@@ -1,9 +1,13 @@
 export { AvatarVideosApiPath } from './enum/enums.js';
 export {
     type AvatarConfig,
+    type Composition,
     type InputKind,
     type RenderAvatarResponseDto,
     type RenderAvatarVideoRequestDto,
+    type Scene,
+    type SceneAvatar,
+    type Script,
     type VideoCodec,
     type VideoFormat,
 } from './types/types.js';
