+++ shared/src/bundles/avatar-videos/types/composition.type.ts
@@ -1,0 +1,9 @@
+import { type Scene } from './scene.type.js';
+import { type Script } from './script.type.js';
+
+type Composition = {
+    scenes: Scene[];
+    scripts: Script[];
+};
+
+export { type Composition };
