+++ shared/src/bundles/avatar-videos/types/scene.type.ts
