+++ shared/src/bundles/avatar-videos/types/script.type.ts
@@ -1,0 +1,9 @@
+type Script = {
+    id: string;
+    duration: number;
+    text: string;
+    voiceName: string;
+    url?: string;
+};
+
+export { type Script };
