+++ shared/src/bundles/avatar-videos/types/types.ts
@@ -1,6 +1,9 @@
 export { type AvatarConfig } from './avatar-config.type.js';
+export { type Composition } from './composition.type.js';
 export { type InputKind } from './input-kind.type.js';
 export { type RenderAvatarVideoRequestDto } from './render-avatar-request-dto.type.js';
 export { type RenderAvatarResponseDto } from './render-avatar-response-dto.type.js';
+export { type Scene, type SceneAvatar } from './scene.type.js';
+export { type Script } from './script.type.js';
 export { type VideoCodec } from './video-codec.type.js';
 export { type VideoFormat } from './video-format.type.js';
